Allow using a global percentage completion
It's actually quite hard to build a front-end for fwupd at the moment
as you're never sure when the progress bar is going to zip back to 0%
and start all over again.

By creating a FuProgress helper object we can create a truely global
percentage completion of the entire transaction, which was always the
grand plan before things got complicated with composite devices.

Using FuProgress we can encode at the plugin level "the erase takes 50%
of the time, the write takes 40% and the read takes 10%". This means
we can have a progressbar which goes up at a consistent speed, even when
multiple devices are being updated or when there are multiple phases
required.

This does mean a fairly major break to the FuDevice API so that we can
pass around the FuProgress objects (or children there-of), but we just
broken plugin ABI and it's a good time to do this kind of thing.